ds/requirements.txt
streamlit
langchain
openai
faiss-cpu
beautifulsoup4
requests
python-dotenv
